Display monster statuses in ^x
This provides greater accessibility to screen-reader players and an
overview elaboration for players who see "…)" in an entry on the console
monster list. This display is also added to webtiles; while the icons do
duplicate this information, the purpose of ^x is to be clarifying, so
the information is spelled out.