Fix map view console cursor clamping
Missing negation in ca36f614e.  For various reasons, this issue only
came up for extremely wide/tall maps in console, so was a hard bug to
spot, but it was instantiated by sprint_mu. On this map, when fully
explored (e.g.  via `&{`) the effect was to let the cursor move off
screen in map/`X` view.