Update floor for features next to damaging walls (12233)
When casting Frozen Ramparts, the floor tiles for features adjacent to
icy walls were not updated. This is a problem for habitable features
like fountains and open doors, since this indicates to the player that
these tiles wouldn't harm a monster. This also applied to slimy walls,
although it's less common for the player to see this. Update the packed
cell code to properly modify floor tiles.

Shallow and deep water are instances where updating floor tiles wouldn't
help, since the water tile isn't transparent. We may need to make
special "icy water" tile variants to handle this case.