Apply location effects *after* temp changing a feature
ea32a606ab reordered things so that location effects were applied to
monsters as a side effect before the temporary feature change took
effect, i.e. using the about-to-be-replaced feature value. Aside from
not applying any effects of the new feature type to a monster on the
position, this caused weird things like a water nymph standing on a trap
and temporarily flooding it to trigger that trap on the temporary
terrain change. Some example crashes:

http://crawl.berotato.org/crawl/morgue/ardl/crash-ardl-20191231-234514.txt
http://crawl.akrasiac.org/rawdata/elzilcho/crash-elzilcho-20190927-225210.txt

(The crash is when the terrain change reverts, and the shaft's trap info
is gone.)

This commit partially reverts ea32a606ab (sorry) but keeps the wall
check in a more direct way.