Allow clicking outside chat window to defocus
focus-trap has a setting that is supposed to do this, but getting it to
work ran up against what I think is a bug in focus-trap, where the
returnFocusOnDeactivate setting isn't respected when the new target
isn't clickable
(https://github.com/davidtheclark/focus-trap/issues/103). The boolean
expression here is a bit ugly, but I did it that way as a minimal change
from the original logic. I have only a somewhat vague understanding
overall of what is going on with focus in this library, so I hope that
this doesn't have further side-effects.

This library seems to be unmaintained now, I wonder if there are any
alternatives to consider instead?

Hopefully resolves #1243