Simplify targeter classes
The shadow step targeter landing site method doesn't need to be a
virtual method in the targeter base class. For uses in
direction_chooser, we can do a static cast of this pointer to one of
targeter_shadow_step, since we always know when our hit function was
initialized with a targeter_shadow_step instance.