Use the magic mapping glyph for unvisited stairs and teleporters
By default this is not a distinct glyph. Using the magic mapping glyph
for unvisited stairs and teleporters allows visually impaired players to
screen read this information easily which was previously only
communicated by colour.

This does mean the magic mapping glyph has a slight overload for stairs
(unvisited vs. completely unseen) but the two different meanings are
close enough and the information about stair connectivity isn't provided
until the player actually views the stair; this is a strict improvement
for visually impaired players.