Placeholder tiles for cigotuvi's embrace
The item tile is a butchering of the shadow dragon armour with a bit of
the shape hacked off and some unpleasant brown-to-green colors replacing
the red.  The equip tile is just an existing green armour tile that was
kicking around unused (at least as far as I saw), it's not really a
perfect match but it's ok-ish for now. I eagerly await replacements from
a real artist, but for a release anything seems better to me than a TODO
tile...