Don't have allies avoid player line of fire
In 340c9e81, monster AI was improved to avoid LOF between a monster's
allies and their targets. This change had the player considered as an
ally, hence summons in front of a player would try to move out of the
path between the player and a hostile. This is convenient when the
player wants to use a ranged attack, but doesn't work well in the more
common case were allies are used for cover. This commit has the player
not be considered a ranged ally by allied monsters, hence they won't
avoid player LOF, but continues to allow that consideration for fellow
allied monsters.

There's probably no good way to predict when the player wants to make a
ranged attack behind allies, so it's possible that a new 't' command is
the best solution. A way to toggle this LOF-avoidance behaviour through
other 't' commands might need some consideration, if something like this
is implemented.