Restore note about pasting to local tiles seed selection
This was lost in 04990682366a as part of improvements to the seed
selection menu (since no button corresponds to [p] and pasting is more
generally possible in local tiles text input). However, I think it's
still useful for players to have a reminder that pasting is possible at
this screen, and describe the way in which it may not behave as
expected.