Make necro a major sac, and charm a moderate sac
In its present state, charm is not a major sacrifice.  Its most powerful spells are dual-school with Hexes or Necromancy, but being barred from those schools has a larger impact because those schools have further powerful spells, some of which are single-school. It could make sense for charms to be a major sacrifice if/when the school has some heft to it once again.

Necromancy, on the other hand, like conjurations or summoning, can be your school of choice for the whole 3 rune game.  Unlike charms, it does not require a weapon, nor significant investment in other schools.