Fix some target prompts under Fedhas
Fedhas completely protects allied plants from the effects of Airstrike,
Shatter, and tin of tremorstones, but players currently receive
incorrect penance warnings when using these effects. Add the relevant
Fedhas ally checks so that the attack prompt won't be triggered for
these cases.

Fedhas currently doesn't protect allied plants from player Freeze, but
the prompt correctly warns them in this case. Freeze and intentional
melee attacks are the remaining exception where we do apply Fedhas
penance, but there's no reason for these exceptions to exist that I'm
aware of. It would easier to centralize the prompting logic in
bad_attack() if all attempts at harming a Fedhas ally by the player were
prevented. This can be addressed in a future commit, but the fixes in
this commit will be enough for release.