Don't call show_init from viewwindow off-level
The easiest way to trigger this was fixed in 726680df4ae, but
`viewwindow` is still needed to handle local tiles resizing in map mode.
This is a quick fix to just prevent `show_init` altogether while on a
level excursion, but the real fix probably involves better refactoring
as described in the comment.