Adjust pandemonium lord spells
Previously, 75% of pan lords were given spells, and the other 25% had
their base stats buffed.

Add the chance for a single spell to the 25% case, using a new spell
list. This list of unique spells is based on movement: either closing
the gap to the player or messing with their positioning.