Seed stability fixes for heliophobic_arrival_battle_scene
This vault had two issues: (i) in the vault code, the number of rng
calls depended on player species (so could affect the seed differently
depending on whether the player is a corpse species). This is easily
fixed. (ii) the corpse generation code actually generates, places, and
kills a monster, which does all sorts of stuff; but the main impact is
that it will along the way generate weapons depending on species, which
leads to substantial differences in rng rolls per species. This is a
kind of weird and deep gotcha for corpse placers (though it mostly
shouldn't matter in terms of seed stability, unless a vault designer is
trying to place a player-based corpse), so I've just isolated the rng
for corpse generation in a subgenerator.

(As a side note, `; nothing` in a monspec is not helpful here, because
that turns out to generate the monster with gear and then take it
away.)