Try to fix flash + input interactions once and for all (12292)
I've tried previously to do some variant of this (RAII cursor state
restoring) in the past and it's always led to inexplicable crashes, but
this time maybe I've gotten it right. Basically, what was happening in
various ways is that redraw calls would lead to moving the cursor out
from under the input code when it goes to draw the current input string,
and it would try to draw at coordinates that are invalid for the current
region. One way this could happen is by setting a flash on the strong
`warning` which always appears in the annotate prompt (found in
Sentinel's rc file, see discussion in
https://crawl.develz.org/mantis/view.php?id=12292&nbn=3#bugnotes).
Flashing calls viewwindow, which tends to leave the cursor in the mlist.
This commit ensures that viewwindow, flash, and a few other things will
always restore the cursor, and it also slightly improves the cursor
position sanity checking. It addresses several past related crashes,
such as d21a89c53800 (where now I've concluded that setting the cursor
region without moving the cursor is a bad idea, and probably led to
different out of bounds errors down the line).

Also, prevent flashing at the main menu.