Set cursor region on fakecursorxy
This was apparently a major factor in "invalid" cursor positions.
Basically, as far as I can tell, write_char_at leads to changes in
wherex() and wherey(), even without a cgotoxy call, and so all the other
checks will flag the results as invalid unless the region is updated.
This also makes the region value consistent after cursorxy regardless of
options.