Fix zaps being rendered under items/clouds on local tiles (Goratrix)
This issue was caused by zaps being added to the main tile buffer before
the tiles for the foreground (items and clouds). The fix is to pack the
overlay tiles in two stages; although the tiles are in the same buffer,
the order of the tiles within the main buffer is reversed, and the
display order is now correct.

This didn't happen on webtiles because it already used the same two-pass
strategy out of neccessity: see cdd0e279bb for more.