Explicitly set the cursor region to GOTO_CRT for map view
I think this is a better general fix than 899df634c; it's just that
right now only webtiles (that I know of) happens to trigger a
`viewwindow` call during map view while the map is up. But since it
happens as a side effect often, it may be best to be prepared.